alx_travel_app/listings/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import path
from .views import (
    ListingViewSet,
    BookingViewSet,
    InitiatePaymentView,
    VerifyPaymentView
)

router = DefaultRouter()
router.register(r'listings', ListingViewSet, basename='listing')
router.register(r'bookings', BookingViewSet, basename='booking')

urlpatterns = router.urls + [
    path('initiate-payment/', InitiatePaymentView.as_view(), name='initiate-payment'),
    path('verify-payment/<str:tx_ref>/', VerifyPaymentView.as_view(), name='verify-payment'),
]
